add tab for hotkeys settings
